fix(commitmsg): multiline trailers and BREAKING CHANGE trailer key
Added support for multiline trailer values. The extra lines must start
with a leaking whitespace. The value is going to be split and then
joined together with simple line feed characters so that the leading
whitespace will not remain in the parsed result.

Added support for marking a commit as having a breaking change in
addition to the '!' after the commit type, flagging it using the trailer
key "BREAKING CHANGE" or "BREAKING-CHANGE".